requirements.txt
openai==0.28.1
streamlit>=1.31.0
pymupdf>=1.23.0
python-docx>=0.8.11
biopython>=1.83
pandas>=2.2.0
python-dotenv>=1.0.1
